+++ backend/src/reverseShare/dto/updateReverseShare.dto.ts
@@ -1,0 +1,6 @@
+import { IsString } from "class-validator";
+
+export class UpdateReverseShareDTO {
+  @IsString()
+  shareExpiration: string;
+}